# [IMP] website, website_sale, web_editor: add `s_floating_blocks` snippet