--- a/addons/website/data/image_library.xml
+++ b/addons/website/data/image_library.xml
@@ -451,6 +451,27 @@
     <field name="type">url</field>
     <field name="url">/website/static/src/img/snippets_demo/s_framed_intro.jpg</field>
 </record>
+<record id="website.s_floating_blocks_1" model="ir.attachment">
+    <field name="public" eval="True"/>
+    <field name="name">s_floating_blocks_1</field>
+    <field name="res_model">ir.ui.view</field>
+    <field name="type">url</field>
+    <field name="url">/website/static/src/img/library/shop_generic_1_lg.jpg</field>
+</record>
+<record id="website.s_floating_blocks_2" model="ir.attachment">
+    <field name="public" eval="True"/>
+    <field name="name">s_floating_blocks_2</field>
+    <field name="res_model">ir.ui.view</field>
+    <field name="type">url</field>
+    <field name="url">/website/static/src/img/library/shop_generic_2_lg.jpg</field>
+</record>
+<record id="website.s_floating_blocks_3" model="ir.attachment">
+    <field name="public" eval="True"/>
+    <field name="name">s_floating_blocks_3</field>
+    <field name="res_model">ir.ui.view</field>
+    <field name="type">url</field>
+    <field name="url">/website/static/src/img/library/shop_generic_3_lg.jpg</field>
+</record>
 <record id="website.s_carousel_intro_default_image_1" model="ir.attachment">
     <field name="public" eval="True"/>
     <field name="name">s_carousel_intro_default_image_1.jpg</field>