--- a/addons/website/static/src/snippets/s_floating_blocks/floating_blocks.edit.scss
+++ b/addons/website/static/src/snippets/s_floating_blocks/floating_blocks.edit.scss
@@ -1,0 +1,43 @@
+// Snippets Modal Preview
+// ======================
+// Performs an animation on :hover showcasing the user how the snippet behaves
+// in use. Measurements and values are arbitrary, the goal is just to provide
+// a preview.
+.o_snippet_preview_wrap {
+    [data-snippet="s_floating_blocks"] {
+        height: 750px;
+
+        // Needed to avoid that inner blocks overflows while translating
+        overflow: hidden;
+
+        .s_floating_blocks_block {
+            opacity: 1;
+
+            &:nth-child(1) {
+                transition: all .3s;
+                margin: 3% 0 -20%;
+                transform-origin: center top;
+                transform: scale(0.98);
+
+                > .s_parallax_bg {
+                    background-position: top !important;
+                }
+            }
+        }
+    }
+    &:hover [data-snippet="s_floating_blocks"] .s_floating_blocks_block:nth-child(1) {
+        margin: 3% 0 -30%;
+        transform: scale(.95);
+    }
+}
+
+// Edit mode
+// =========
+.s_floating_blocks {
+    .s_floating_blocks_block {
+        > .container, > .container-fluid, > .o_container_small {
+            // Ensure that grid items cannot be moved outside their cards.
+            overflow: hidden;
+        }
+    }
+}