--- a/addons/website/views/snippets/snippets.xml
+++ b/addons/website/views/snippets/snippets.xml
@@ -30,9 +30,9 @@
                     t-thumbnail="/website/static/src/img/snippets_thumbs/s_instagram_page.svg"/>
                 <t t-if="debug" snippet-group="debug" t-snippet="website.s_snippet_group" string="Debug"
                     t-thumbnail="/website/static/src/img/snippets_thumbs/s_debug_group.png"/>
-                <t id="installed_snippets_hook"/>
-                <t snippet-group="products" string="Products" t-install="website_sale"
+                <t snippet-group="products" t-snippet="website.s_snippet_group" string="Products"
                     t-thumbnail="/website/static/src/img/snippets_thumbs/s_dynamic_products.svg"/>
+                <t id="installed_snippets_hook"/>
                 <t snippet-group="blogs" string="Blogs" t-install="website_blog"
                     t-thumbnail="/website/static/src/img/snippets_thumbs/s_blog_posts.svg"/>
                 <t snippet-group="events" string="Events" t-install="website_event"
@@ -504,6 +504,8 @@
             t-image-preview="/website/static/src/img/snippets_previews/s_newsletter_sms_notifications_preview.png"/>
         <t t-install="website_payment" string="Donation" group="contact_and_forms"
             t-image-preview="/website/static/src/img/snippets_previews/s_donation_preview.png"/>
+        <t t-install="website_sale" string="Products Carousel" group="products"
+            t-image-preview="/website/static/src/img/snippets_previews/s_dynamic_snippet_products_preview.jpg"/>
     </xpath>
     <xpath expr="//snippets[@id='snippet_content']" position="inside">
         <t id="newsletter_snippet" t-install="mass_mailing" string="Newsletter" t-thumbnail="/website/static/src/img/snippets_thumbs/s_newsletter_subscribe_form.svg"/>