--- a/addons/website_sale/views/snippets/snippets.xml
+++ b/addons/website_sale/views/snippets/snippets.xml
@@ -2,10 +2,6 @@
 <odoo>
 
 <template id="snippets" inherit_id="website.snippets" name="e-commerce snippets">
-    <xpath expr="//t[@id='installed_snippets_hook']" position="after">
-        <t snippet-group="products" t-snippet="website.s_snippet_group" string="Products"
-            t-thumbnail="/website/static/src/img/snippets_thumbs/s_dynamic_products.svg"/>
-    </xpath>
     <xpath expr="//t[@id='sale_products_hook']" position="replace">
         <t t-snippet="website_sale.s_dynamic_snippet_products" string="Products" group="products"/>
     </xpath>