… addons/website/static/src/img/library/shop_generic_1_lg.jpg
… addons/website/static/src/img/library/shop_generic_2_lg.jpg
… addons/website/static/src/img/library/shop_generic_3_lg.jpg
… addons/website/static/src/img/snippets_previews/s_dynamic_snippet_products_preview.jpg
… addons/website/views/snippets/s_floating_blocks.xml