This PR introduce a new static snippet tailored for eCommerce categories and promos.

It also make the "Products" snippets category visible by default, allowing the injection of snippets that don't require `website_sale` to be installed.

task-4373543
part-of-task-4252024

Related:
- https://github.com/odoo/design-themes/pull/1054

| Screenshoot |
|--------|
| <img width="1340" alt="image" src="https://github.com/user-attachments/assets/f037f3df-5528-4372-9108-6b7d3eb5b49e"> | 



---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr