@stefanorigano 2025-01-03
Hello @sobo-odoo 👋🙂
Could you have a look at this? Already approved by BOJE/LEBL. 

Note: we'll enable the "Products" category by default, even if `website_sale` it's not installed (similarly to newsletter snippets).

Thanks!