@sobo-odoo 2025-01-03
@stefanorigano Wow this snippet is really cool! :scream::star_struck: I don't think I can review it unfortunately, as I am working on the refactoring and given that the PR is huge :grimacing:  Maybe ask BVR or QSM (or someone else) on Monday instead :slightly_smiling_face: