@stefanorigano 2025-01-03
Thanks for the fast feedback @sobo-odoo 👍 
Ok, @bvr-odoo or @qsm-odoo  then.

Thanks in advance!