@stefanorigano 2025-02-27
Hello @qsm-odoo , I've updated the code. 

About my issue with `this._refreshPublicWidgets();` (sorry, I must have deleted my message by mistake), I've ultimately solved it  [here](https://github.com/odoo/odoo/pull/189299/files#diff-9691e7243f8efab4b4b0512510f6c54786ca7fceca9579b07325fd5af3e14e05R85) using `this.trigger_up('content_changed')`.

Let me know if you're up for review or if we should delegate.

Thanks