@stefanorigano 2025-03-29
Hello @qsm-odoo , thanks for your review.
I've updated the code accordingly, "resolved" the most obvious comments and replied where necessary.

In therms of git history, at the current state we have 3 commits:
- Enable the dialog's  "Products" tab for static snippets
- Add the new snippet
- Add a glob pattern for snippet' scss in edit mode (for the modal preview).

Thanks.