@qsm-odoo 2025-03-31
Trying out the snippet with looking at the code now, it looks cool.

I have this quirky overlay placement:
![image](https://github.com/user-attachments/assets/a99c63c4-918f-4c52-8bbe-7d6ef515d5b7)

It can happen in two cases:
- Clicking on a card, then simply scrolling a bit. The clicked card is "still there", then the overlay still hovers it. Maybe we could fix that on the overlay code itself (trying to check if the element is not "behind another" but not sure it would be easy/reliable (you can be behind and be visible)). But maybe there is a way to avoid that issue on snippet level @stefanorigano ? Maybe forcing hidden card to be actually display: none (in which case, the overlay would disappear)? That would actually make the transparent case a bit better, making fully covered cards disappear :shrug: 
![image](https://github.com/user-attachments/assets/60227926-8995-4ee9-9a13-c1bd522ed4d0)

- Adding a new card: when you click the button, we scroll to the new card... but we keep the old one enabled. Should not we either not scroll to the new item or enable the new item (I don't think enabling the new item is consistent with our other behaviors... but that's the one I would choose: it's weird to see option of cards on the right but those are about the old cards, not the new one). We could also scroll to the new item but disabling the old overlay :shrug: 

cc @lebl-odoo