@stefanorigano 2025-04-02
> it looks cool.

@qsm-odoo , thanks and thanks  for your review.

> I have this quirky overlay placement:

Ooo thanks for noticing it... during the last iterations I've oversimplified the code and removed by mistake one part making cards to stack with a slight gap.  Edition was confusing in the scenario you described indeed.

It has been fixed and, when it's the case, it's now clear that the card selected is NOT the one on top.

| desktop | mobile |
|--------|--------|
| ![image](https://github.com/user-attachments/assets/ed4f20c7-8386-481c-b882-a38ef7acdc8c) | ![image](https://github.com/user-attachments/assets/fdeb9114-c154-48f4-83dc-3902e31068f7) | 

---

> Adding a new card: when you click the button, we scroll to the new card... but we keep the old one enabled.[...]

Right, thanks. Solved by activating the newly created card.

---

// Other chnages

- Make use of CSS `contain` and `backface-visibility` to prove performance
- Decreased `dynamicContent` workload. `opacity` is now set during initialization, `top` when the header changes only.
- Use cached values to check if _isInViewport rather than using getBoundingClientRect
- Pre-compute transform values
- Fixed some missing some `El` and `Els` suffixes on variables name.

---

The diff -> https://github.com/odoo/odoo/compare/f8cea4138d45a260bf43c63292706af12fcd575d..8f3786a29ff820d076b619c83acb20f60db7f71b

---

Thanks!