@stefanorigano 2025-04-04
> performance. 

 I registered a diminution of slow frames from  1% to 3% , depending by the test.
Intel/OSX, Firefox (no cache). The page contained  5 `s_floating_blocks` snippets, each one with a video and 2 parallax cards.

<details><summary>Details</summary>
<p>

```
===== BEFORE ======

===== PERFORMANCE TEST RESULTS ===== debugger eval code:155:17
Test type: fast scroll debugger eval code:156:17
Duration: 10.03 seconds debugger eval code:157:17
Total scroll events: 314 debugger eval code:158:17
Average scroll handler time: 0.04ms debugger eval code:159:17
Frame count: 314 debugger eval code:160:17
Average FPS: 31.2 debugger eval code:161:17
Slow frames (>16.67ms): 237 (75.7%) debugger eval code:162:17
Very slow frames (>50ms): 64 (20.4%) debugger eval code:163:17
==================================== debugger eval code:164:17

 ===== PERFORMANCE TEST RESULTS ===== debugger eval code:155:17
Test type: slow scroll debugger eval code:156:17
Duration: 10.008 seconds debugger eval code:157:17
Total scroll events: 341 debugger eval code:158:17
Average scroll handler time: 0.11ms debugger eval code:159:17
Frame count: 341 debugger eval code:160:17
Average FPS: 34.0 debugger eval code:161:17
Slow frames (>16.67ms): 293 (86.2%) debugger eval code:162:17
Very slow frames (>50ms): 32 (9.4%) debugger eval code:163:17
==================================== debugger eval code:164:17

===== PERFORMANCE TEST RESULTS ===== debugger eval code:155:17
Test type: erratic scroll debugger eval code:156:17
Duration: 10.003 seconds debugger eval code:157:17
Total scroll events: 126 debugger eval code:158:17
Average scroll handler time: 0.20ms debugger eval code:159:17
Frame count: 404 debugger eval code:160:17
Average FPS: 40.3 debugger eval code:161:17
Slow frames (>16.67ms): 258 (64.0%) debugger eval code:162:17
Very slow frames (>50ms): 44 (10.9%) debugger eval code:163:17
==================================== debugger eval code:164:17


​===== AFTER ======

===== PERFORMANCE TEST RESULTS ===== debugger eval code:155:17
Test type: fast scroll debugger eval code:156:17
Duration: 10.013 seconds debugger eval code:157:17
Total scroll events: 308 debugger eval code:158:17
Average scroll handler time: 0.07ms debugger eval code:159:17
Frame count: 308 debugger eval code:160:17
Average FPS: 30.7 debugger eval code:161:17
Slow frames (>16.67ms): 223 (72.6%) debugger eval code:162:17
Very slow frames (>50ms): 71 (23.1%) debugger eval code:163:17
==================================== debugger eval code:164:17

 ===== PERFORMANCE TEST RESULTS ===== debugger eval code:155:17
Test type: slow scroll debugger eval code:156:17
Duration: 10.009 seconds debugger eval code:157:17
Total scroll events: 325 debugger eval code:158:17
Average scroll handler time: 0.08ms debugger eval code:159:17
Frame count: 325 debugger eval code:160:17
Average FPS: 32.4 debugger eval code:161:17
Slow frames (>16.67ms): 283 (87.3%) debugger eval code:162:17
Very slow frames (>50ms): 26 (8.0%) debugger eval code:163:17
==================================== debugger eval code:164:17

===== PERFORMANCE TEST RESULTS ===== debugger eval code:155:17
Test type: erratic scroll debugger eval code:156:17
Duration: 10.071 seconds debugger eval code:157:17
Total scroll events: 129 debugger eval code:158:17
Average scroll handler time: 0.11ms debugger eval code:159:17
Frame count: 412 debugger eval code:160:17
Average FPS: 40.8 debugger eval code:161:17
Slow frames (>16.67ms): 262 (63.7%) debugger eval code:162:17
Very slow frames (>50ms): 34 (8.3%) debugger eval code:163:17
==================================== debugger eval code:164:17

```

</p>
</details> 


> "While this interaction is alive, the opacity is supposed to be 1". The moment and the way it happens is irrelevant.

Clear, thanks for the explanation ( @mano-odoo , interesting hint for your stuff 😉 )


> Grid Stuff -> Maybe it's a Firefox issue? 🤷 Did not tested elsewhere.

Nah, I'm on Firefox as well and I can reproduce your issue by just using a different screen, it's matter of viewport size indeed, uff. 
My hack enforce the number of rows, but to so that I have to enforce the number of columns as well. I could fix that but I'm sure that another issue would popup eventually. 
We have to get rid of the hack indeed, you're right. Pushed a new tmp on top 

Thanks